feat: refactor player input handling using separate UserInput component
- Removed PlayerMovement.cs and migrated all movement logic into Player.cs
- Created UserInput.cs to centralize input detection (WASD + Jump)
- Player.cs now queries global UserInput via FindObjectOfType
- Movement logic is cleaner and separated from input detection
