--- Assets/Scripts/Player.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: d3e8be3d68e7744fe86f10f74373946b
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
