--- Assets/Scripts/Player/PlayerMovement.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 5a67025a9413e4a2eb0af62dccef5a3a
