+++ Assets/Scripts/UserInput.cs
@@ -1,0 +1,20 @@
+using UnityEngine;
+
+public class UserInput : MonoBehaviour
+{
+    public bool Left { get; private set; }
+    public bool Right { get; private set; }
+    public bool Jump { get; private set; }
+
+    void Update()
+    {
+        Left = Input.GetKey(KeyCode.A);
+        Right = Input.GetKey(KeyCode.D);
+        Jump = Input.GetKeyDown(KeyCode.Space);
+    }
+
+    public void ResetJump()
+    {
+        Jump = false;
+    }
+}
